Address comments in #99
