+++ src/utils/getConfigManager.ts
@@ -3,36 +3,45 @@ import * as path from 'path';
 import ConfigSet from '../types/ConfigSet';
 import getOptions from './getOptions';
 import loadConfig from './loadConfig';
-import { getConfigPathCandidatesForFile } from './getConfigsForFile';
+import { getConfigPathCandidatesForFile } from './getConfigPathCandidatesForFile';
 import NormalizedPath from '../types/NormalizedPath';
 
 class ConfigManager {
-    private configSet: ConfigSet = null;
-
+    private fullConfigSet: ConfigSet = null;
     // The subset of configs that has been loaded
     private partialDiscoveredConfigs: ConfigSet = {};
 
     // The set of paths we have checked for configs in the filesystem
     private discoveredPaths: Set<string> = new Set();
 
-    public get all(): ConfigSet {
-        if (this.configSet === null) {
+    public getAllConfigs(): ConfigSet {
+        if (this.fullConfigSet === null) {
             this._getAllConfigs();
         }
-        return this.configSet;
+        return this.fullConfigSet;
     }
 
-    public partialConfigSetForPath(configSourcePath: NormalizedPath): ConfigSet {
+    public getPartialConfigSetForPath(configSourcePath: NormalizedPath): ConfigSet {
         const partialSet: ConfigSet = {};
 
-        if (this.configSet) {
-            for (let configPathCandidate of getConfigPathCandidatesForFile(configSourcePath)) {
-                if (this.configSet[configPathCandidate]) {
-                    partialSet[configPathCandidate] = this.configSet[configPathCandidate];
+        const configCandidatesForFile = getConfigPathCandidatesForFile(configSourcePath);
+
+        if (this.fullConfigSet) {
+            // If the full config set has been initialized (e.g. by calling cfgManager.getAllConfigs)
+            // then instead of doing redundant fs access, construct the result from the full config
+            // set
+            for (let configPathCandidate of configCandidatesForFile) {
+                if (this.fullConfigSet[configPathCandidate]) {
+                    partialSet[configPathCandidate] = this.fullConfigSet[configPathCandidate];
                 }
             }
         } else {
-            for (let configPathCandidate of getConfigPathCandidatesForFile(configSourcePath)) {
+            // If the full config set has not been initialized, go to disk to find configs in the
+            // candidate set.
+            //
+            // As we scan paths, we add them to our partial configs and our set of checked paths
+            // so we can avoid redudnant fs access for this same fence and path in the future.
+            for (let configPathCandidate of configCandidatesForFile) {
                 const configPathCandidateFull = path.join(configPathCandidate, 'fence.json');
                 if (this.discoveredPaths.has(configPathCandidateFull)) {
                     const discoveredConfig = this.partialDiscoveredConfigs[configPathCandidate];
@@ -58,7 +67,7 @@ class ConfigManager {
     }
 
     private _getAllConfigs() {
-        this.configSet = {};
+        this.fullConfigSet = {};
 
         let files: string[] = [];
         for (let rootDir of getOptions().rootDir) {
@@ -66,7 +75,7 @@ class ConfigManager {
         }
 
         files.forEach(file => {
-            loadConfig(file, this.configSet);
+            loadConfig(file, this.fullConfigSet);
         });
     }
 }
